# Improve AST clone performance