--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var isArray = require('isarray');
-var objectKeys = require('object-keys');
+var objectKeys = Object.keys || require('object-keys');
 var indexOf = require('indexof');
 var reduce = require('array-reduce');