--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -21,17 +21,15 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
     }
 
     function cloneContainer (collector, obj) {
-        if (isArray(obj)) {
-            return cloneArray(collector, obj);
-        } else if (obj.type && whitelist[obj.type]) {
+        if (obj.type && whitelist[obj.type]) {
             return cloneNode(collector, obj);
         } else {
             return cloneObj(collector, obj);
         }
     }
 
-    function cloneArray (collector, ary) {
-        var i, len;
+    function cloneArray (ary) {
+        var i, len, collector = [];
         for (i = 0, len = ary.length; i < len; i += 1) {
             collector.push(cloneOf(ary[i]));
         }
@@ -64,8 +62,10 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         if (typeof val === 'object' && val !== null) {
             if (val instanceof RegExp) {
                 return new RegExp(val);
+            } else if (isArray(val)) {
+                return cloneArray(val);
             } else {
-                return cloneContainer(isArray(val) ? [] : {}, val);
+                return cloneContainer({}, val);
             }
         } else {
             return val;