--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -17,14 +17,14 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
     }, {});
 
     function cloneRoot (ast) {
-        return cloneContainer({}, ast);
+        return cloneContainer(ast);
     }
 
-    function cloneContainer (collector, obj) {
+    function cloneContainer (obj) {
         if (obj.type && whitelist[obj.type]) {
-            return cloneNode(collector, obj);
+            return cloneNode(obj);
         } else {
-            return cloneObj(collector, obj);
+            return cloneObj(obj);
         }
     }
 
@@ -36,16 +36,16 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         return collector;
     }
 
-    function cloneNode (collector, obj) {
-        var i, len, props = whitelist[obj.type];
+    function cloneNode (obj) {
+        var i, len, props = whitelist[obj.type], collector = {};
         for (i = 0, len = props.length; i < len; i += 1) {
             cloneProperty(collector, obj, props[i]);
         }
         return collector;
     }
 
-    function cloneObj (collector, obj) {
-        var key;
+    function cloneObj (obj) {
+        var key, collector = {};
         for (key in obj) {
             cloneProperty(collector, obj, key);
         }
@@ -65,7 +65,7 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
             } else if (isArray(val)) {
                 return cloneArray(val);
             } else {
-                return cloneContainer({}, val);
+                return cloneContainer(val);
             }
         } else {
             return val;