--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -16,10 +16,6 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         return props;
     }, {});
 
-    function cloneRoot (ast) {
-        return cloneContainer(ast);
-    }
-
     function cloneContainer (obj) {
         if (obj.type && whitelist[obj.type]) {
             return cloneNode(obj);
@@ -72,5 +68,5 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         }
     }
 
-    return cloneRoot;
+    return cloneContainer;
 };