--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -18,9 +18,9 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
 
     function cloneContainer (obj) {
         if (obj.type && whitelist[obj.type]) {
-            return cloneNode(obj);
+            return cloneProps(obj, whitelist[obj.type]);
         } else {
-            return cloneObj(obj);
+            return cloneProps(obj, objectKeys(obj));
         }
     }
 
@@ -32,18 +32,10 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         return collector;
     }
 
-    function cloneNode (obj) {
-        var i, len, props = whitelist[obj.type], collector = {};
+    function cloneProps (obj, props) {
+        var i, len, collector = {};
         for (i = 0, len = props.length; i < len; i += 1) {
             cloneProperty(collector, obj, props[i]);
-        }
-        return collector;
-    }
-
-    function cloneObj (obj) {
-        var key, collector = {};
-        for (key in obj) {
-            cloneProperty(collector, obj, key);
         }
         return collector;
     }