--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -17,8 +17,9 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
     }, {});
 
     function cloneContainer (obj) {
-        if (obj.type && whitelist[obj.type]) {
-            return cloneProps(obj, whitelist[obj.type]);
+        var props = obj.type ? whitelist[obj.type] : null;
+        if (props) {
+            return cloneProps(obj, props);
         } else {
             return cloneProps(obj, objectKeys(obj));
         }