--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -34,17 +34,14 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
     }
 
     function cloneProps (obj, props) {
-        var i, len, collector = {};
+        var i, len, key, collector = {};
         for (i = 0, len = props.length; i < len; i += 1) {
-            cloneProperty(collector, obj, props[i]);
+            key = props[i];
+            if (obj.hasOwnProperty(key)) {
+                collector[key] = cloneOf(obj[key]);
+            }
         }
         return collector;
-    }
-
-    function cloneProperty (collector, obj, key) {
-        if (obj.hasOwnProperty(key)) {
-            collector[key] = cloneOf(obj[key]);
-        }
     }
 
     function cloneOf (val) {