--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -26,9 +26,9 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
     }
 
     function cloneArray (ary) {
-        var i, len, collector = [];
-        for (i = 0, len = ary.length; i < len; i += 1) {
-            collector.push(cloneOf(ary[i]));
+        var i = ary.length, collector = [];
+        while (i--) {
+            collector[i] = cloneOf(ary[i]);
         }
         return collector;
     }