--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -7,12 +7,9 @@ var reduce = require('array-reduce');
 
 module.exports = function cloneWithWhitelist (astWhiteList) {
     var whitelist = reduce(objectKeys(astWhiteList), function (props, key) {
-        var currentProps = astWhiteList[key];
-        var prepend = [];
-        if (indexOf(currentProps, 'type') === -1) {
-            prepend.push('type');
-        }
-        props[key] = prepend.concat(currentProps);
+        var propNames = astWhiteList[key];
+        var prepend = (indexOf(propNames, 'type') === -1) ? ['type'] : [];
+        props[key] = prepend.concat(propNames);
         return props;
     }, {});