--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -13,7 +13,7 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         return props;
     }, {});
 
-    function cloneContainer (obj) {
+    function cloneObject (obj) {
         var props = obj.type ? whitelist[obj.type] : null;
         if (props) {
             return cloneProps(obj, props);
@@ -48,12 +48,12 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
             } else if (isArray(val)) {
                 return cloneArray(val);
             } else {
-                return cloneContainer(val);
+                return cloneObject(val);
             }
         } else {
             return val;
         }
     }
 
-    return cloneContainer;
+    return cloneObject;
 };