--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -13,12 +13,12 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         return props;
     }, {});
 
-    function cloneObject (obj) {
+    function cloneNodeOrObject (obj) {
         var props = obj.type ? whitelist[obj.type] : null;
         if (props) {
-            return cloneProps(obj, props);
+            return cloneNode(obj, props);
         } else {
-            return cloneProps(obj, objectKeys(obj));
+            return cloneObject(obj);
         }
     }
 
@@ -30,13 +30,23 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         return collector;
     }
 
-    function cloneProps (obj, props) {
+    function cloneNode (obj, props) {
         var i, len, key, collector = {};
         for (i = 0, len = props.length; i < len; i += 1) {
             key = props[i];
             if (obj.hasOwnProperty(key)) {
                 collector[key] = cloneOf(obj[key]);
             }
+        }
+        return collector;
+    }
+
+    function cloneObject (obj) {
+        var props = objectKeys(obj);
+        var i, len, key, collector = {};
+        for (i = 0, len = props.length; i < len; i += 1) {
+            key = props[i];
+            collector[key] = cloneOf(obj[key]);
         }
         return collector;
     }
@@ -48,12 +58,12 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
             } else if (isArray(val)) {
                 return cloneArray(val);
             } else {
-                return cloneObject(val);
+                return cloneNodeOrObject(val);
             }
         } else {
             return val;
         }
     }
 
-    return cloneObject;
+    return cloneNodeOrObject;
 };