--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -22,33 +22,33 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         }
     }
 
-    function cloneArray (ary) {
-        var i = ary.length, collector = [];
-        while (i--) {
-            collector[i] = cloneOf(ary[i]);
-        }
-        return collector;
-    }
-
-    function cloneNode (obj, props) {
-        var i, len, key, collector = {};
+    function cloneNode (node, props) {
+        var i, len, key, clone = {};
         for (i = 0, len = props.length; i < len; i += 1) {
             key = props[i];
-            if (obj.hasOwnProperty(key)) {
-                collector[key] = cloneOf(obj[key]);
+            if (node.hasOwnProperty(key)) {
+                clone[key] = cloneOf(node[key]);
             }
         }
-        return collector;
+        return clone;
     }
 
     function cloneObject (obj) {
         var props = objectKeys(obj);
-        var i, len, key, collector = {};
+        var i, len, key, clone = {};
         for (i = 0, len = props.length; i < len; i += 1) {
             key = props[i];
-            collector[key] = cloneOf(obj[key]);
+            clone[key] = cloneOf(obj[key]);
         }
-        return collector;
+        return clone;
+    }
+
+    function cloneArray (ary) {
+        var i = ary.length, clone = [];
+        while (i--) {
+            clone[i] = cloneOf(ary[i]);
+        }
+        return clone;
     }
 
     function cloneOf (val) {