--- a/lib/clone-ast.js
+++ b/lib/clone-ast.js
@@ -22,6 +22,14 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         }
     }
 
+    function cloneArray (ary) {
+        var i = ary.length, clone = [];
+        while (i--) {
+            clone[i] = cloneOf(ary[i]);
+        }
+        return clone;
+    }
+
     function cloneNode (node, props) {
         var i, len, key, clone = {};
         for (i = 0, len = props.length; i < len; i += 1) {
@@ -39,14 +47,6 @@ module.exports = function cloneWithWhitelist (astWhiteList) {
         for (i = 0, len = props.length; i < len; i += 1) {
             key = props[i];
             clone[key] = cloneOf(obj[key]);
-        }
-        return clone;
-    }
-
-    function cloneArray (ary) {
-        var i = ary.length, clone = [];
-        while (i--) {
-            clone[i] = cloneOf(ary[i]);
         }
         return clone;
     }