--- a/lib/create-whitelist.js
+++ b/lib/create-whitelist.js
@@ -1,11 +1,11 @@
 'use strict';
 
 var defaultProps = require('./ast-properties');
-var objectKeys = require('object-keys');
-var extend = require('xtend');
+var objectKeys = Object.keys || require('object-keys');
+var assign = require('object-assign');
 
 module.exports = function createWhitelist (options) {
-    var opts = extend({}, options);
+    var opts = assign({}, options);
     var typeName, i, len;
     var keys = objectKeys(defaultProps);
     var result = {};