--- a/lib/create-whitelist.js
+++ b/lib/create-whitelist.js
@@ -11,7 +11,7 @@ module.exports = function createWhitelist (options) {
     var result = {};
     for (i = 0, len = keys.length; i < len; i += 1) {
         typeName = keys[i];
-        result[typeName] = [].concat(defaultProps[typeName]).concat(opts.extra);
+        result[typeName] = defaultProps[typeName].concat(opts.extra);
     }
     return result;
 };