--- a/package.json
+++ b/package.json
@@ -12,8 +12,8 @@
     "array-reduce": "0.0.0",
     "indexof": "0.0.1",
     "isarray": "^1.0.0",
-    "object-keys": "^1.0.4",
-    "xtend": "^4.0.0"
+    "object-assign": "^4.0.1",
+    "object-keys": "^1.0.4"
   },
   "devDependencies": {
     "babel-types": "^6.3.20",