Trying to improve AST clone performance more and more.

```
$ node bench_deep_tree.js
clone x 1.59 ops/sec ±1.18% (8 runs sampled)
1.1.0 x 14.56 ops/sec ±3.46% (38 runs sampled)
1.5.0 x 483 ops/sec ±1.12% (85 runs sampled)
head  x 525 ops/sec ±1.06% (86 runs sampled)
Fastest is head
$ 
```

- 8.7% faster than previous version (espurify 1.5.0)
- 36x faster than espurify 1.1.0
- 330x faster than npm clone module!

by [Benchmark.js](https://benchmarkjs.com/)